Refactor alert & fix tests & improve README
Move the `alert.js` file to `aws-alert/index.js`
Fix error in findScriptPath ('function' => 'faas')
Fix error in configure (directories must exist to write files into them)
Improve the README by:
  1. fixing the link to the function's define file
  2. update function constraints and timing decision
