+++ lib/faas/aws-alert/index.js
