+++ lib/faas/handler.js
@@ -1,6 +1,6 @@
 /* eslint-disable no-underscore-dangle */
 
-const alert = require('./alert')
+const alert = require('./aws-alert')
 const func = require('./aws-func')
 const task = require('./task')
