+++ lib/index.js
@@ -18,8 +18,12 @@ const constants = {
   CompatibleServerlessSemver: '^1.0.3',
   DefaultScriptName: 'script.yml',
   ServerlessFile: 'serverless.yml', // duplicated below but mentioned here for consistent use
+  ServerlessDirectories: [
+    'aws-alert',
+    'aws-func',
+  ],
   ServerlessFiles: [
-    'alert.js',
+    'aws-alert/index.js',
     'handler.js',
     'aws-func/index.js', 'aws-func/define.js', 'aws-func/invoke.js', 'aws-func/handle.js', 'aws-func/valid.js',
     'package.json', 'package-lock.json',
@@ -71,7 +75,7 @@ const impl = {
       if (impl.fileExists(localDefaultScript)) {
         return localDefaultScript
       } else {
-        return path.join(__dirname, 'function', constants.DefaultScriptName)
+        return path.join(__dirname, 'faas', constants.DefaultScriptName)
       }
     }
   },
@@ -597,6 +601,10 @@ module.exports = {
       if (options.debug) {
         console.log('No conflicts found, creating a local copy of the function assets for deployment.')
       }
+      constants.ServerlessDirectories.forEach((dir) => {
+        const destPath = path.join(cwd, dir)
+        fs.mkdirSync(destPath)
+      })
       constants.ServerlessFiles.forEach((file) => {
         const sourcePath = path.join(__dirname, 'faas', file)
         const destPath = path.join(cwd, file)
